drivers: modem: Add support for commands that don't have a line ending
Some commands need to be processed before a "\r\n" is available and
there might also be commands that have "\r\n" as data but doesn't mean
the end of the command.

To solve this a MODEM_CMD_DIRECT has been added. cmd_handler_process()
will look for matching direct commands before checking if a whole line
is available for matching the normal commands.

A direct command can use modem_cmd_handler_data.ret to return a value to
cmd_handler_process. This value can be either -EAGAIN, meaning that more
data is needed or it will return the number of bytes to skip forward, ie
the length of the command that was handled.

Also, modem_cmd_handler_data.eol was also added to be able to handle
devices with different line-endings for commands sent to the modem
device.

Signed-off-by: Tobias Svehagen <tobias.svehagen@gmail.com>